requirements.txt
click==6.7
Flask==0.12.2
gevent==1.2.2
greenlet==0.4.12
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
python-dateutil==2.6.1
six==1.11.0
Werkzeug==0.12.2
